Added tests for 2xx series and 5xx series
