+++ integration/http-compliance-tests/src/test/java/org/wso2/carbon/gateway/httpcompliance/tests/responses/servererror/HTTP501ComplianceTest.java
@@ -1,4 +1,20 @@
-package org.wso2.carbon.gateway.httpcompliance.tests.successful;
+/*
+ * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.wso2.carbon.gateway.httpcompliance.tests.responses.servererror;
 
 import io.netty.handler.codec.http.HttpMethod;
 import io.netty.handler.codec.http.HttpResponseStatus;
@@ -21,12 +37,12 @@ import static org.wso2.gw.emulator.http.server.contexts.HttpServerConfigBuilderC
 import static org.wso2.gw.emulator.http.server.contexts.HttpServerRequestBuilderContext.request;
 import static org.wso2.gw.emulator.http.server.contexts.HttpServerResponseBuilderContext.response;
 
-public class HTTP204ComplianceTest {
+public class HTTP501ComplianceTest {
     private GatewayAdminClient gwClient;
     private HttpServerOperationBuilderContext emulator;
-    private static final String HOST = "127.0.0.1";
+    private String host = "127.0.0.1";
     private int port = 9090;
-    private static final String RESPONSE_BODY = "";
+    private String servererror = "Method not implemented";
 
     @BeforeClass
     public void setup() throws Exception {
@@ -41,19 +57,11 @@ public class HTTP204ComplianceTest {
         return Emulator.getHttpEmulator().server().given(configure().host("127.0.0.1").port(6065).context("/users"))
 
                 .when(request()
-                        .withMethod(HttpMethod.POST)
+                        .withMethod(HttpMethod.OPTIONS)
                         .withPath("/user1"))
                 .then(response()
-                        .withStatusCode(HttpResponseStatus.NO_CONTENT)
-                        .withBody(RESPONSE_BODY))
-
-                .when(request()
-                        .withMethod(HttpMethod.POST)
-                        .withPath("/user1")
-                        .withBody("Body included"))
-                .then(response()
-                        .withStatusCode(HttpResponseStatus.NO_CONTENT)
-                        .withBody(RESPONSE_BODY))
+                        .withStatusCode(HttpResponseStatus.NOT_IMPLEMENTED)
+                        .withBody(servererror))
 
                 .operation().start();
     }
@@ -66,37 +74,20 @@ public class HTTP204ComplianceTest {
     }
 
     @Test
-    public void test204POSTRequestWithoutBody() throws Exception {
+    public void test500OPTIONSRequest() throws Exception {
         HttpClientResponseProcessorContext response = Emulator.getHttpEmulator().client()
-                .given(HttpClientConfigBuilderContext.configure().host(HOST).port(port))
+                .given(HttpClientConfigBuilderContext.configure().host("127.0.0.1").port(9090))
 
                 .when(HttpClientRequestBuilderContext.request()
+                        .withMethod(HttpMethod.OPTIONS)
                         .withPath("/new-route")
-                        .withMethod(HttpMethod.POST)
                         .withHeader("routeId", "r1"))
 
                 .then(HttpClientResponseBuilderContext.response().assertionIgnore()).operation().send();
 
-        Assert.assertEquals(response.getReceivedResponse().getStatus(), HttpResponseStatus.NO_CONTENT,
+        Assert.assertEquals(response.getReceivedResponse().getStatus(), HttpResponseStatus.NOT_IMPLEMENTED,
                 "Expected response code not found");
-        Assert.assertNull(response.getReceivedResponseContext().getResponseBody());
-    }
 
-    @Test
-    public void test204POSTRequestWithBody() throws Exception {
-        HttpClientResponseProcessorContext response = Emulator.getHttpEmulator().client()
-                .given(HttpClientConfigBuilderContext.configure().host(HOST).port(port))
-
-                .when(HttpClientRequestBuilderContext.request()
-                        .withPath("/new-route")
-                        .withMethod(HttpMethod.POST)
-                        .withBody("Body included")
-                        .withHeader("routeId", "r1"))
-
-                .then(HttpClientResponseBuilderContext.response().assertionIgnore()).operation().send();
-
-        Assert.assertEquals(response.getReceivedResponse().getStatus(), HttpResponseStatus.NO_CONTENT,
-                "Expected response code not found");
-        Assert.assertNull(response.getReceivedResponseContext().getResponseBody());
+        Assert.assertEquals(response.getReceivedResponseContext().getResponseBody(), servererror);
     }
 }
