+++ integration/http-compliance-tests/src/test/java/org/wso2/carbon/gateway/httpcompliance/tests/responses/successful/HTTP202ComplianceTest.java
@@ -1,4 +1,4 @@
-package org.wso2.carbon.gateway.httpcompliance.tests.successful;
+package org.wso2.carbon.gateway.httpcompliance.tests.responses.successful;
 
 import io.netty.handler.codec.http.HttpMethod;
 import io.netty.handler.codec.http.HttpResponseStatus;
@@ -21,16 +21,18 @@ import static org.wso2.gw.emulator.http.server.contexts.HttpServerConfigBuilderC
 import static org.wso2.gw.emulator.http.server.contexts.HttpServerRequestBuilderContext.request;
 import static org.wso2.gw.emulator.http.server.contexts.HttpServerResponseBuilderContext.response;
 
-public class HTTP205ComplianceTest {
+public class HTTP202ComplianceTest {
     private GatewayAdminClient gwClient;
     private HttpServerOperationBuilderContext emulator;
     private static final String HOST = "127.0.0.1";
     private int port = 9090;
-    private static final String REQUEST_BODY = "Body included";
-    private static final String RESPONSE_BODY = "";
+    private static final String RESPONSE_BODY = "Request accepted and will be processed in due time";
+    private File clientRequestPayload;
 
     @BeforeClass
     public void setup() throws Exception {
+        clientRequestPayload = new File(getClass().getClassLoader()
+                .getResource("test-payloads" + File.separator + "client-request-payload.txt").toURI());
         gwClient = new GatewayAdminClientImpl();
         gwClient.startGateway();
         gwClient.deployArtifact("artifacts" + File.separator + "http-compliance-test-camel-context.xml");
@@ -39,21 +41,23 @@ public class HTTP205ComplianceTest {
     }
 
     private HttpServerOperationBuilderContext startHttpEmulator() {
-        return Emulator.getHttpEmulator().server().given(configure().host("127.0.0.1").port(6065).context("/users"))
+        return Emulator.getHttpEmulator().server().given(configure().host(HOST).port(6065).context("/users"))
 
                 .when(request()
+                        .withPath("/user1")
                         .withMethod(HttpMethod.POST)
-                        .withPath("/user1"))
+                        .withBody(clientRequestPayload))
                 .then(response()
-                        .withStatusCode(HttpResponseStatus.RESET_CONTENT)
+                        .withStatusCode(HttpResponseStatus.ACCEPTED)
                         .withBody(RESPONSE_BODY))
 
                 .when(request()
-                        .withMethod(HttpMethod.POST)
                         .withPath("/user1")
-                        .withBody(REQUEST_BODY))
+                        .withMethod(HttpMethod.POST)
+                        .withBody(clientRequestPayload)
+                        .withBody("Body included"))
                 .then(response()
-                        .withStatusCode(HttpResponseStatus.RESET_CONTENT)
+                        .withStatusCode(HttpResponseStatus.ACCEPTED)
                         .withBody(RESPONSE_BODY))
 
                 .operation().start();
@@ -67,39 +71,41 @@ public class HTTP205ComplianceTest {
     }
 
     @Test
-    public void test205POSTRequestWithoutBody() throws Exception {
+    public void test202POSTRequestWithPayloadWithoutBody() throws Exception {
         HttpClientResponseProcessorContext response = Emulator.getHttpEmulator().client()
                 .given(HttpClientConfigBuilderContext.configure().host(HOST).port(port))
 
                 .when(HttpClientRequestBuilderContext.request()
                         .withMethod(HttpMethod.POST)
+                        .withPath("/new-route")
                         .withHeader("routeId", "r1")
-                        .withPath("/new-route"))
+                        .withBody(clientRequestPayload))
 
                 .then(HttpClientResponseBuilderContext.response().assertionIgnore()).operation().send();
 
-        Assert.assertEquals(response.getReceivedResponse().getStatus(), HttpResponseStatus.RESET_CONTENT,
+        Assert.assertEquals(response.getReceivedResponse().getStatus(), HttpResponseStatus.ACCEPTED,
                 "Expected response code not found");
 
-        Assert.assertNull(response.getReceivedResponseContext().getResponseBody());
+        Assert.assertEquals(response.getReceivedResponseContext().getResponseBody(), RESPONSE_BODY);
     }
 
     @Test
-    public void test205POSTRequestWithBody() throws Exception {
+    public void test202POSTRequestWithPayloadWithBody() throws Exception {
         HttpClientResponseProcessorContext response = Emulator.getHttpEmulator().client()
                 .given(HttpClientConfigBuilderContext.configure().host(HOST).port(port))
 
                 .when(HttpClientRequestBuilderContext.request()
-                        .withPath("/new-route")
                         .withMethod(HttpMethod.POST)
+                        .withPath("/new-route")
                         .withHeader("routeId", "r1")
-                        .withBody(REQUEST_BODY))
+                        .withBody(clientRequestPayload)
+                        .withBody("Body included"))
 
                 .then(HttpClientResponseBuilderContext.response().assertionIgnore()).operation().send();
 
-        Assert.assertEquals(response.getReceivedResponse().getStatus(), HttpResponseStatus.RESET_CONTENT,
+        Assert.assertEquals(response.getReceivedResponse().getStatus(), HttpResponseStatus.ACCEPTED,
                 "Expected response code not found");
 
-        Assert.assertNull(response.getReceivedResponseContext().getResponseBody());
+        Assert.assertEquals(response.getReceivedResponseContext().getResponseBody(), RESPONSE_BODY);
     }
 }
